ceil is still a float. cast to int

+++ spectral/_spectral.py
@@ -314,7 +314,7 @@ class Spectral(object):
 
     def stft(self, sig):
         s = np.pad(sig, (self.wlen//2, 0), 'constant')
-        cols = np.ceil((s.shape[0] - self.wlen) / self.fshift + 1)
+        cols = int(np.ceil((s.shape[0] - self.wlen) / self.fshift + 1))
         s = np.pad(s, (0, self.wlen), 'constant')
         frames = as_strided(s, shape=(cols, self.wlen),
                             strides=(s.strides[0]*self.fshift,
